Add one example of rspec test and move web-console to dev only.
I don't know why the web console was complaining but it wanted to be moved to
dev only which made sense. I also added one very basic controller spec.

Add Simplecov

Simplecov is a helpful tool for tracking test coverage. It can be made to
automatically fail a build or test run if the test coverage is not high enough
or drops too much in any one build. I added it here in order to track coverage
and set extremely low cuttoffs. As we add more tests we can raise the cuttoffs
to match. Alternatively set the cuttoff high and force ourselves to work up to it.